Move DailyTourPromo into the same dir as the CardGrid
which is the only place it's used
