+++ common/views/components/CardGrid/CardGrid.tsx
@@ -1,7 +1,7 @@
 import { classNames, cssGrid } from '../../../utils/classnames';
 import ExhibitionPromo from '../ExhibitionPromo/ExhibitionPromo';
 import EventPromo from '../EventPromo/EventPromo';
-import DailyTourPromo from '../DailyTourPromo/DailyTourPromo';
+import DailyTourPromo from './DailyTourPromo';
 import BookPromo from '../BookPromo/BookPromo';
 import Layout12 from '../Layout12/Layout12';
 import StoryPromo from '../StoryPromo/StoryPromo';
